+++ daos/queries/post/getById.sql
@@ -6,8 +6,8 @@ SELECT p.id
      , p.linked_bill_id
      , p.view_count
      , p.is_deleted
-     , TO_CHAR(p.created_at AT TIME ZONE 'Asia/Seoul', 'YYYY-MM-DD HH24:MI') AS created_at
-     , TO_CHAR(p.updated_at AT TIME ZONE 'Asia/Seoul', 'YYYY-MM-DD HH24:MI') AS updated_at
+     , TO_CHAR(p.created_at, 'YYYY-MM-DD HH24:MI') AS created_at
+     , TO_CHAR(p.updated_at, 'YYYY-MM-DD HH24:MI') AS updated_at
      , u.nickname
      , u.provider
      , b.bill_name        AS linked_bill_name